Now in GetDiasSemanaOrdenados the first param is DiaSemana

+++ Turnos.Model/Entities/UsuarioTurnoSemanalSecuenciaDia.cs
@@ -20,7 +20,7 @@ namespace Turnos.Model.Entities
 
         public override string ToString()
         {
-            var primerDiaSemana = (int)UsuarioTurnoSemanalSecuencia.UsuarioTurnoSemanal.Usuario.Empresa.PrimerDiaSemana;
+            var primerDiaSemana = UsuarioTurnoSemanalSecuencia.UsuarioTurnoSemanal.Usuario.Empresa.PrimerDiaSemana;
 
             var diasSemana = Utilities.GetDiasSemanaOrdenados(primerDiaSemana);
             return String.Format("Dia {0} ({1}) Turno: {2}.", Dia, diasSemana.ElementAt(Dia - 1), Turno.ToString());
